# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Keeparr API",
    "version": "0.3.6",
    "description": "Keeparr tags what to keep and reports what can be deleted — it never deletes anything itself.\n\n**Authentication**: most endpoints use the `keeparr_session` cookie (set by logging in through the web UI). Endpoints marked with `X-Api-Key` security also accept the admin API key (Settings → General) in the `X-Api-Key` header — intended for automation (`/api/stats`, `/api/admin/jobs`).\n\nAll ids (`ratingKey`) are the media server's item ids (Plex ratingKey or Jellyfin/Emby item id)."
  },
  "servers": [{ "url": "/", "description": "This Keeparr instance" }],
...
  "components": {
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "keeparr_session",
        "description": "Signed session cookie set by the web login."
      },
      "apiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Api-Key",
        "description": "Admin API key from Settings → General."
      }
    },
...
  },
  "security": [{ "cookieAuth": [] }],
  "paths": {
...
    "/api/keep": {
      "post": {
        "tags": ["media"],
        "summary": "Keep a title (this user). Clears their don't-care + OK-to-delete atomically.",
        "requestBody": { "content": { "application/json": { "schema": { "type": "object", "properties": { "ratingKey": { "type": "string" } }, "required": ["ratingKey"] } } } },
        "responses": {
          "200": { "description": "OK" },
          "404": { "description": "`{error:'unknown_item'}` — unknown or removed from the server" }
        }
      },
      "delete": {
        "tags": ["media"],
        "summary": "Remove this user's keep (others' keeps stay)",
        "requestBody": { "content": { "application/json": { "schema": { "type": "object", "properties": { "ratingKey": { "type": "string" } }, "required": ["ratingKey"] } } } },
        "responses": { "200": { "description": "OK" } }
      }
    },
    "/api/skip": {
      "post": {
        "tags": ["media"],
        "summary": "Mark \"don't care\" (this user). Clears their keep + OK-to-delete atomically.",
        "requestBody": { "content": { "application/json": { "schema": { "type": "object", "properties": { "ratingKey": { "type": "string" } }, "required": ["ratingKey"] } } } },
        "responses": {
          "200": { "description": "OK" },
          "404": { "description": "`{error:'unknown_item'}` — unknown or removed from the server" }
        }
      },
      "delete": {
        "tags": ["media"],
        "summary": "Clear this user's \"don't care\"",
        "requestBody": { "content": { "application/json": { "schema": { "type": "object", "properties": { "ratingKey": { "type": "string" } }, "required": ["ratingKey"] } } } },
        "responses": { "200": { "description": "OK" } }
      }
    },
    "/api/mark-delete": {
      "post": {
        "tags": ["media"],
        "summary": "Mark \"OK to delete\" — only on titles this user requested via Seerr",
        "requestBody": { "content": { "application/json": { "schema": { "type": "object", "properties": { "ratingKey": { "type": "string" } }, "required": ["ratingKey"] } } } },
        "responses": {
          "200": { "description": "OK" },
          "403": { "description": "`{error:'not_requested'}` — not this user's request" },
          "404": { "description": "`{error:'unknown_item'}` — unknown or removed from the server" }
        }
      },
      "delete": {
        "tags": ["media"],
        "summary": "Clear this user's \"OK to delete\"",
        "requestBody": { "content": { "application/json": { "schema": { "type": "object", "properties": { "ratingKey": { "type": "string" } }, "required": ["ratingKey"] } } } },
        "responses": { "200": { "description": "OK" } }
      }
    },
    "/api/skip-batch": {
      "post": {
        "tags": ["media"],
        "summary": "Skip a batch (\"don't care\" all) and deal a fresh one",
        "description": "Skipping clears this user's keep / OK-to-delete on the batch (mutually exclusive). Unknown/tombstoned keys are ignored. Max 500 keys.",
        "requestBody": { "content": { "application/json": { "schema": { "type": "object", "properties": { "ratingKeys": { "type": "array", "items": { "type": "string" }, "maxItems": 500 } } } } } },
        "responses": {
          "200": { "description": "`{items: MediaCard[], remaining}`" },
          "400": { "description": "`{error:'bad_request'}` (not an array) or `{error:'too_many_items'}` (over 500)" }
        }
      }
    },
...
  }
}